Initial commit
This is a very old project. I built it when I was just learning
Android and had no idea of version control system, code
style/smell, testing etc.

This is working source code and optimised to reduce apk size
while packing in as many features as possible. While doing so,
code readability is compromised. But I hope it is
understandable for someone having little experience developing
Android apps, knowledge of Android Studio features and some
patience.